Jour 1 - LogEntry add/count/view
Commit cumulatif (oubli de commit à chaque fonctionalité)
Possible d'entrer un nouveau logEntry
Compte le nombre d'entrées -> modification de ce que la méthode retourne int to long
affiche les log sauf pour les planêtes proches et peut-être pas dans le bon sens
ajout d'une méthode pour récupérer les images dans la DB Barkeley

+++ src/ca/qc/cvm/dba/scoutlog/app/Facade.java
@@ -43,7 +43,7 @@ public class Facade {
 		return app.getPlanetList();
 	}
 	
-	public int getNumberOfEntries() {
+	public long getNumberOfEntries() {
 		return app.getNumberOfEntries();
 	}
